refactor(comfyui): 翻訳のレスポンス解析を共有モジュールへ寄せます
Google翻訳のレスポンス解析が`translate-text`と`anime-video-quick`で重複していて、
このPRで両者の方針が分岐していました。
`translate-text`は`non_empty_list`を新設して不正な要素を`ValueError`で弾く形にしたのに対し、
`anime-video-quick`は`cast`と`continue`で不正な要素を黙って読み飛ばす形のままです。
同じ外部APIの同じ形を2箇所で別々に解釈していると、
片方だけ直した時に挙動差が残ります。

`share_encode.py`と同じ仕組みに乗せて、
`custom-node/translate.py`へ1箇所に寄せます。
リクエストの組み立てまで含めた`translate_to_english`を公開して、
空の翻訳結果も`ValueError`にします。
翻訳できなかった時に原文へ倒すかどうかは呼び出し側の判断なので、
共有モジュール側では倒さずに落とします。

symlinkは`writeCheckedNode`の`builtins.readDir`が拾うので、
`custom-node.nix`側の変更は要りません。

両ノードのderivationに`translate.py`が実体として並ぶことと、
壊れた形のレスポンスがそれぞれのメッセージで落ちることを確認しました。
