+++ nixos/host/bullet/comfyui/custom-node/anime-video-quick/translate.py
@@ -1,0 +1,1 @@
+../translate.py
