+++ nixos/host/bullet/comfyui/custom-node/translate-text/translate.py
@@ -1,0 +1,1 @@
+../translate.py
